Switch Num from veneer to private macro interface
